The local clock is not working in tandem with all other clocks (doesn't stop when user starts editing a clock and doesn't change value when user tries to convert times).
This is because, so far we have been treating the default local clock and
the different clocks added by the user as two separate entities. The
primary motivation to keep the local clock separate was our original
plan to give local clock some special treatment (like not letting the user
delete it, not bubbling it up in the list of clocks etc.)

However, we now realize that keeping the local clock separate makes the
code messy because all operations should be done on the local clock AND
the user added clocks.

So, in this commit, all clocks (including the local clock) are put in the
same array. The separation is achieved by some index manipulation in the
TimeZoneClocksManager and binding DOM elements only to a section of array
etc.