Prior to this commit, offset based clocks were handled incorrectly. Offset based clocks (like UTC-7 should always show UTC-7 irrespective of the current day light savings scenario. However, because we were mapping an offset like UTC-7 to a regular time zone like America/Los_Angeles, the actual offset used by clock varies based on the current day light savings situation.
To remedy this, we start treating absolute offset based clocks
completely separate from location based clocks. To that end, major
changes were made to the way data is consumed, a new offset based
time zone object is created and a major refactor around the code
that adjusts clocks when one of the clocks is edited were made.