So far the code has been running on a hack that refreshed the view every second (a dummy $interval in the main controller). In this commit, we remove the hack and instead update the view whenever a timestamp changes in the clocks or manually invoke $apply whenever we respond to user actions and make changes that need to be reflected on the UI.
There was also a bug which where we wouldn't show 'Invalid time' warning
if user entered an invalid time and clicked away (instead of hitting
return). Turns out using agngular's GetterSetter option is causing the
problem although am not sure why. For now, to workaround that problem
I stopped using the gettersetter methodolody and instead directly bound
the timePart to the view.